Adding PropEr tests for crossbar bindings.
The tests work by first generating a pattern, and then using it to
expand it based on the rules that * means exactly 1 part of the pattern
is a wildcard and that # means 0..infinity patterns can be matched. The
test tries to generate both good and bad cases to run over the match
function and make sure it works fine.

At the time of this commit, the function is failing

Note: amended to change to a better distribution.