v1.56: catch when no master account can be determined and return the unmodified json object.
On clean installations, when creating the first account, there is obviously no master account id yet. This causes a crash in wh_reseller when setting the pvt fields on the new account. Chicken/egg problem. So catch the crash, log it, and continue with the unmodified json object.

Perhaps it should finishing saving the first account and retry the setting of service plans?