KAZOO-46: handle unicode in json strings
Sometimes clients send JSON in containing unicode characters, encoded as
Latin-1 strings, which the json decoders aren't crazy about at all. It expects
the JSON to be UTF-8 binaries. When decoding the JSON, catch exceptions related
to the decoder failing on the Latin-1 encoding, and use the unicode module to
convert the JSON from Latin-1 to UTF-8 and retry the decode.