KAZOO-487: fetch mailbox based on calling device's owner settings
If the owner_id has 0 voicemail boxes, prompt for box number
If the owner_id has 1 voicemail box, prompt for the box's pin
If the owner_id has 2 or more voicemail boxes, search the boxes'
"mailbox" property to see if one matches the caller's
caller-id-number. If so, use that box and prompt for the pin; otherwise
prompt for a box number.