Support for call confirmation on non forwarded calls
Freeswitch requires group_* values to be in the global variables of the bridge to work.

Until now, these values were only in CCVS because it was for forwarded calls only. It worked because it uses a loopback.

If we want to be able to do it on regular bridges, this change is needed.