Revert "start metaflows for external caller"
This reverts commit d28918f249ed73dc64e9e13b7472bedbbfccb22b.

This commit set the autorizing ID to the account id as a means to start konami for the account.  However, the unintended side-effect is the authorizing id is published by whapps_call and therefore present on the CDR as well as subsequent requests.  External apps and some Kazoo components use the presence of authorizing id to denote an on-net call.

The prefered fix would be in konami, to make a utility function that will return the authorizing id or account id if not authorized.  This new function should be used anywhere the authorizing id is needed.