local_render: create system_alert render modules
rather than using the pool, where a process has to build the render
module then render the template, create fixed renderer modules, based on
the content-types found in the system_alert notification, to render the
payloads.

This should significantly reduce the processing needed for
system_alerts, as well as free the render farm up for other types of
notifications.

It could be a good idea to extend this thought to all the templates,
build their system_config versions statically, and only use the render
farm for account-overridden templates