Fixed an issue with adding a customer
When someone needed to add a customer without having their credit card braintree send an error that make_default(undefined) clause is not defined in braintree_card.erl. now we can add a new customer in braintree by just having using his info and adding a card is not needed.