Modernize calls to random-generating functions
In Erlang/OTP 19 the 'random' module is marked deprecated and the 'rand'
module is expected to be used instead.  This module also exists in 18.x.

Also, the compiler issues warnings for crypto:rand_bytes/1 and prompts for
the use of crypto:strong_rand_bytes/1 instead.

This set of changes allows the kazoo code base to be compiled without errors
in 19.0 (because the current compiler options consider warnings as errors)
but most likely disables support for 17.x (because the 'rand' module is not
present there).

It should be merged to 'master' *only* if support for 17.x is not desirable.
However, if support for 17.x is still needed, the code should take out the
compiler option that turns warnings into errors, so that the code can also be
compiled in 19.0.