patch-emer-cid: set emergency cid on gateway(s)
When setting caller id, gateways were blindly setting outbound caller id
from the JSON payload (which could be the external caller id). When the
endpoints were formatted, the "global" caller id was the emergency CID
while each gateway's CID was the external CID. Since external and
emergency CID didn't match, the gateway CID was not stripped in
stepswitch_util:format_endpoints/4.

This patch sets the gateway CID to the emergency CID if
appropriate. Formatting will then operate on the emergency CID and be
stripped if unchanged (no formatting rules on the resource).