HELP-26829: do not compute cascade quantities twice (#3041)
* HELP-26829: do not compute cascade quantities twice

It seems we're applying cascade quantities twice during
calculating services charges.

Currently in `kz_services:calclute_services_charges/2` gets current
quantities, apply cascade quantities and then set the result (which
is current quanitties + updates + Cascade quantities) into the service jobj which
looks okay at the first sight.

But in the line below that, when the items are created for the updated
service jobj, `kz_service_plans:create_items/2` is calling `kz_service_plans:create_items/1`
which by default is calculating cascade quantities.

This results to adding cascade quatinties to updated service twice.

In this patch I remove the call to `apply_cascade_quantities/2`, because
in the end when creating service from json, it will calculate anyway.
This still needs to be tested though.

* HELP-26829: better logging