Alter system configs REST API (diffs, validation)
A system configuration document is a JSON document of certain
syntetic type (each node value must be a JSON object of type
system_config.$config_name). REST API allows to address both
documents and and docuemnt nodes. Default node of the
configuration document is treated specially, see below.

In order to avoid confusion configuration document nodes
we call document sections, each such section is a JSON
object representing a configuration object for Kazoo
node or zone.

General idea is to return (with with_defaults=true URI
option) the effective values for configuration, therefore
a complex mergedown is performed: document values are
merged with document default section, and then merged
with schema defaults.

Therefore on save actions (PUT/POST/PATCH) a reverse
operation is performed: a difference from defaults
is calculated and stored to the document (if any).
Thus by setting a value to schema default will effectively
remove it from the document on save action.

Documents
---------

GET always returns either an empty document or stored
document. The document always has a default section included,
populated with default values deduced from schema.

POST expects to receive a complete configuration document.
The document must be valid, each section of the document
must pass the validation against system_config.$config_name
schema. Before saving the difference with provided default
is calculated, and then the difference with schema defaults
is calculated, and then the result of this is stored, meaning
if section value is equal to default section value it will
not be stored to the document. The actual (stored) values
are then returned as POST results.

PATCH will merge provided changes with stored document,
and then the resulting document is valid then the same
store procedure is applied as in POST. The actual stored
values are then returned as PATCH result.

DELETE deletes the document (or wipes a section). Attempt
to delete non-existing configuration object will generate
an HTTP 404 error.

Sections
--------

GET always returns either an empty document or stored
values.

POST expects to receive a valid configuration document
(against schema system_config.$config_name). If the
document is valid, then a difference is calculated
with configuration document default section, and
then the difference is calculated with schema defaults,
and then the result is stored. For default section
only the difference from schema defaults are calculated.
The actual (stored) values are returned as result of
POST request.

PATCH will merge provided changes with stored section,
and then the result is validated against schema. If
result is valid, then the same procedure is applied
as in POST. The actual (stored) values are returned
as PATCH result.

DELETE deletes the section specified. If configuration
document doesn't exist (or secion in the document), then
the request will fail with HTTP 404.

Defaults
--------

In order to have effective configuration values one must
provide `with_defaults=true` URI parameter for GET requests
(either to get document or document section). With this
parameter provided a configuration mergedown as described
above is performed.