HELP-30196 conference control should not rely on destroy event
conference control was waiting for conference-destroy event but if there
are two ecallmgrs, one of them is not going to get the event because of
de-duplication.

since the server was not destroyed, it was getting the commands for the
conference if another instance came alive, but the instanceId is
different and the command was not processed.

a better way to solve this would be to change cb_conferences to issue
the commands to conferences based on the InstanceId (requires changing
the routing key for commands in kapi_conference) and have conference
control bind to InstanceId or Conference.InstanceId