Added option to enable fail_on_single_reject
When forwarding call or during call failover, stepswitch creates bridge request with multiple endpoints which causes freeswitch to call again and again on the same device even if the call is rejected, this will allow to reject the call if one of the user's device cancel's the call. This does not effect group call.