Enable override of the 'make -j' flag (#4420)
`make -j` allows `make` to spawn as many parallel jobs as it needs when
building a large number of modules. The problem with this is that it
can, and does, overwhelm systems with restricted memory, CPU power, or
`ulimit` settings, and it can (and does) cause builds on such systems to
fail, which consumes an enormous amount of time when doing multiple test
runs during development or maintenance.

This commit adds the `make` variable `MAKEFLAGS`, which defaults to `-j`.
This variable can be overridden on a case-to-case basis, to limit or
eliminate parallel jobs.

The change made is backwards-compatible with existing build invocations.

`make all MAKEFLAGS=""     # Eliminate parallel jobs`
`make all MAKEFLAGS="-j8"  # Limit parallel jobs to a max of 8`