Removed special calendar arithmetics for years starting on Monday
If I understand correctly, iso_week_to_gregorian_date({Y,W}) should
return the starting day (Monday) of the ISO week; but when the year Y
starts on Monday, it computes the starting day of the _next_ week.

The change could be terribly broken if its purpose is not what I
guessed.