HELP-34519: don't do after-notify action if notify was not successful
Check for any completed in the notification response and if it wasn't
successful don't do nothing.

This for the case when the user don't have any email address set or
doesn't want to recieve notification or the template is disabled in
account but `delete_after_notify` or `save_after_notify` is set on
voicemail box, so kazoo_voicemail would do that action regardless
of the reality that it wasn't successful.