log a warning when nodes of the same name have mismatched versions
Since status checks are common, it can be helpful to see if/when a
node is running a different version from the rest. In this case, we
use the 'name' portion of the node name to determine what nodes should
match versions. So kazoo_apps, ecallmgr, kamailio, all will compare
versions properly.