HELP-37054-4.2: address ownership of conference dial channels (#4866)
* HELP-37054-4.2: address ownership of conference dial channels

When there are multiple ecallmgrs connected to a FreeSWITCH server, it
is possible that the conference dial occurs on ecallmgr1 and the fetch
request for the loopback-b channel comes in on ecallmgr2. This results
in the legs associated with the loopback (a and b) and the final
b-leg, are all considered "local" to ecallmgr2.

However, the conference dial needs to start a call control process on
ecallmgr1 to get a control queue for the add_participant API (used to
start a conf_participant process to manage the now-dialed
participant).

This means both ecallmgr1 and ecallmgr2 consider the b-leg to be
"local". Metaflows checks the local ecallmgr's channel cache to
determine if it should repond to channel API actions - since both
ecallmgrs consider the channel local, both metaflow listeners publish
the payloads for konami to execute.

In this particular case, two CHANNEL_PIVOTs are picked up by the
cf_exe process and cause the conference action to be skipped (since
both "set_variables" actions call cf_exe:continue/1), the callflow is
exhausted, and cf_exe tears down the channel.

* HELP-37054-4.2: don't recurse into the Flow JSON object

The channels API should allow folks to include "terminators":[] for
appropriate callflow actions. With the default
"remove_recursive":true, the terminators key is removed; when
callflows executes the action, it sees the now-missing terminators key
and includes the default list (the exact opposite of what was
requested).

* HELP-37054-4.2: don't recurse to remove empties

metaflows uses kapi_route:publish_req/1 to include the metaflow (which
may contain empty terminators). Don't recursively strip empty values.

* HELP-37054-4.2: don't recurse on CHANNEL_PIVOT events

CHANNEL_PIVOT events will contain the flow, which may contain empty
terminators, which we don't want to strip.