HELP-37225: don't consider [] as empty
Especially since tts/play commands consider [] different from
'undefined' for terminators. It is getting hard to know which API
commands might have a callflow payload (or a terminators attribute in
general); easier to permit a key with [] when it make no difference
than stripping keys with [] when there is a difference.