HELP-37171-4.2: check if media is convertible to prompt://
When overriding prompts with other forms of media, like tone_stream or
HTTP URLs, normalization still occurs as it would for prompts like
"/{ACCOUNT_ID}/{PROMPT}" or "/system_media/{PROMPT}" or
"{PROMPT}". This results in invalid "prompt://" URLs and FreeSWITCH
being unable to play the appropriate media back to the call leg.

Introduce a function that looks for media that isn't
prompt-normalizable and returns a boolean indicating such.