Fix rate import race condition
Prior to this PR, tasks would execute each row in the CSV using the provided task module, mark the task as complete, and then run the task's cleanup if provided.

The kt_rates used cleanup to do the actual uploading of rates into the ratedecks. Due to the tight testing in pqc_cb_rates:seq/0, the test would query for the rate before the task's cleanup function had finished uploading, resulting in a failed test (subject to race conditions).

This PR introduces a finish callback for tasks to run after all CSV rows are processed but before the task is marked as complete.