don't flush a missing ETS table
If the cache hasn't been started and a `flush` is attempted, a crash
will occur when performing the ETS operation. This can happen when the
Kazoo application hasn't yet started and an operator attempts to flush
its cache.

Instead, check that the cache is running and log if not, avoiding a
crash.