HELP-40835: allow client to opt-out of validation checks
Backport of #5400 to 4.3

When creating a storage plan, Kazoo checks the settings by storing a
small text file with the attachment's settings and then reading it
back. This typically works fine for "dumb" storage backends.

However, it could be a hinderance when the storage backend is "smart"
but only for the specific files it expects to receive. Adding code to
handle the writing and reading of the test file is burdensome to the
storage backend. Presumably once the backend is receiving files
properly, executing more validation checks (for new account storage
plans) could be a waste of time.

This change allows the skipping of the validation of the attachment's
settings during the API request; only JSON and internal KAZOO
validation will be performed.

* add note on dangers of not validating settings

* fix doc formatting

* sysadmins must opt-in to allowing clients to skip validation

* put config in kzs_plan for others to check

* fix validation error

* helpers to toggle validating settings

* update docs on how to allow validation skip

* test skipping validation

* new json schema