HELP-44744: allow chaining children off temporal route actions (#5864)
Previously, actions of `menu`, `enable`, `disable`, and `reset` would
terminate the call after processing. Upon discussion, there wasn't a
compelling reason to continue doing this (and restricting the ability
to chain more actions off the action).

The usecase is that when there is an "emergency" a callflow can be
dialed to disable the main temporal route rules and enable the
particular "emergency" temporal route rule.