[4.3] KZ00-5: move call to cast (#5896)
To avoid blocking the calling process when the gen_server is overly
busy, move the channel updates to a cast instead of a call. In
particular, we've observed the authz worker (during route req/resp for
dialplan) getting blocked for 2s while waiting for this update to
occur.

Since the authz worker and dialplan response don't depend on the
channel cache, no sense blocking for the ETS write.