KAZOO-6118: strip missing connections off storage plan settings
Given a storage plan with "plan".Classification."types".Type:
If Type."connection" is defined, check that the value (the connection
ID) is also defined on the root level's "connections" object. If not,
remove the connection from the Type.

Adds tests for crossbar rejection of undefined connections and checks
that kzs_plan:plan/1 returns an appropriately filtered plan.