HELP-45446: adding hangup cause and disposition to command error event
Adding the hangup reason to command execution error event can help the caller to
determine if it need to continue the processing or not.

For example cf_device can check if the bridge command failed because of hangup
and if yes, then stop the callflow right there, so cf_exe won't have any race
condition of continuing to the next callflow and receiving channel_destroy event.