FORUM-10598: validate using parent account's media in sub accounts
Specifically the ask was to use a parent account's media in a child
account's menu greeting. pqc_forum_10598.erl effectively follows the
path of creating a media file in a parent account, creating a menu in
a child account, and following the path of that
media path (/parent_account/media_id) until it is resolved to a couch
URL by ecallmgr via media_mgr.

Also added a CRUD seq test for the menu API.