Allow generic branching of callflow based on capture group
Most callflow actions that include a feature code option have a
separate module for handling the capture group portion, and a main
action module for the feature.

Rather than require all actions wanting to be feature-code-able to
create a corresponding feature code module, when the action only wants
to use the capture group to re-hunt and branch to the found callflow,
this callflow action branch_bnumber can be put at the end to perform
the hunt.