add from_list_recursive/2 so we don't convert list to binary
from_list_recursive tends to convert a list if all lists' elements are ascii or
printable. Empty list is consider as printable too.

This will cause list values in objects like `[{<<"key">>, []}]` be convert to
binary which should not be!
This commit adds options to hint the from_list_recursive that the object doesn't
have any string (prinatable chars) in list and list should be considered as
list!