PISTON-953: no need to flush endpoint cache after doc updates
Flush should only flush the endpoint cache and callers of the fun do not
need to do so as they are performing a doc update anyway. Recent changes
to kazoo_caches remove this need.

This also resolves an infinite loop where a listener might act on a
doc_{created,deleted,edited} event and request a flush of the endpoint
cache before requesting the endpoint via kz_endpoint:get