note when collection times out
formatting

bring the parent's call-id along

when a federated listener is spun up, it defaulted to all 0s for the
log ID, making it hard to correlate the listener to the parent
gen_listener.

Add the zone tag to the call-id of the parent to help associate the
listener but also differentiate it from other federated listeners.

associate the UUID of the conference event with the logs

use accessor

accessors