Protect system when pagination is disabled (#6162)
* Protect system when pagination is disabled

When fetching large datasets with pagination turned off, it is
possible to tank the system when the dataset causes memory pressure
during loading.

This change adds an optional system config parameter for determining
the maximum memory allowed for the Erlang process handling the API
request.

Additionally, even requests with `pagination=false` (or /v1) will
still be loaded on a per-page basis. Each page load will check the
memory consumption and error with an HTTP 416 response if memory usage
exceeds the configured maximum.

Until the maximum is reached, however, the pages will continue to be
fetched until exhausted as before.

FixtureDB has an added utility to make mirroring an existing CouchDB
database to the FixtureDB fixtures directory for easier conversion of
testing data.