Monitor connection PID before deleting generated files
When using sendfile in Crossbar, there can exist a race condition
where the filename has been passed to Cowboy for sending, Cowboy
handles streaming the file asynchronously, and calls Crossbar's
finish_request callback.

Because the streaming and finish_request callback are happening in
parallel, it is possible for the file to be deleted from the local
disk before streaming has completed.

pqc_cb_ledgers will occassionally provoke this race condition.

Instead, monitor the connection PID until it exits (the TCP session is
finished) and then cleanup the file (cap the wait to 5 minutes just in
case).