[4.3] PROD-170: Announcement Only Mode for vmboxes (#6349)
Announcement Only Mode for Vmboxes

This PR adds new functionality to a voicemail box, enabling a mailbox
to only play the greeting and advance to the next node of the current
callflow.

Announcement only mode can be controlled with the `announcement_only`
boolean in the vmbox schema or by the mailbox owner via the voicemail
config menu.  This default setting for this feature's is disabled or
`false`.

Mailbox owners will now hear new options in the voicemail
configuration menu as two new prompts have been added to
[Kazoo Core Media](https://github.com/2600hz/kazoo-sounds/tree/master/kazoo-core#importing-prompts-for-a-language).

Two new en-US prompts have been added to handle this feature and
should be imported in conjunction with the upgrade of KAZOO.

* vm-settings\_menu
* vm-settings\_menu\_announcement_on

For instructions on how to import prompts, please consult the [Kazoo Core Media documentation](https://github.com/2600hz/kazoo-sounds/tree/master/kazoo-core#importing-prompts-for-a-language).