[4.3] KZOO-57: calculate presence id for route requests (#6405)
* [4.3] KZOO-57: calculate presence id for route requests

When a device registers, the calculated presence_id is cached and
passed along on subsequent INVITEs. If a user hotdesks into the
device, the cached presence_id (likely the SIP username of the device)
will continue to be used on call events related to that device.

This causes issues with applications (like Qubicle) that track a
user's calls based on the user's presence_id being present on dialog
update events (confirmed and terminated).

Instead, provide a uniform way to calculate a device's presence_id and
use it for setting CCVs on the route response in callflow and in
registrar for authn responses.