when validating couchdb view, give the function a name
We use CouchDB `couchjs` to validate that view are a valid Javascript code
and has no syntax error. Start with CouchDB 3, `couchjs` is changed now
it really require that we pass it a valid function statement with name.

This adds a name to function before writing it to file for validation.