[4.3] KCRO-24: address pagination issue when querying MODBs (#6519)
Prior, when page_size and filters are in play with a query, the
next_db clause would not match (since PageSize wasn't
'infinity'). This would lead to results being missed in the intial
query and subsequent queries would skip over to the next group of
results.

Instead, consider if there is a next last key only for whether to
query the current DB again.