[4.3] HELP-14697: correct erroneous data in CDR (#6534)
`duration`, `billsec`, `progresssec` and their ilk should always be
non-negative.

A scenario occurred where a call leg was queued to originate in
FreeSWITCH while an immediate hangup API call was made over
mod_kazoo. The channel was still in `CS_NEW` and went from `DOWN` ->
`HANGUP` without ever being in `UP`.

The result was `variable_duration` being a negative UNIX timestamp and
throwing off all the other calculations (like Ringing-Seconds) in the
CDR that rely on `duration`.

This commit ensures that the variables with relative time are never
negative.