And also it looks like when we originate a call from Kazoo and "park" the call FS bridges the call to loopback so the test ecallmg_fs_channel:is_bridged() returns true even though it is only bridged to itself.
When we then play a message to the Agent it causes ecallmgr_call_command:play() to use "broadcast" rather than "playback".

And therefore the CHANNEL_EXECUTE_COMPLETE "FILE PLAYED" event is again blocked and instead only the "BROADCAST SENT" is received
but this is received before the file is played and can't be used as a trigger to continue the call setup.

This fixes the problem by ignoring bridge to loopback but may not be the right way to fix it.