[4.3] Upstream 4.3 telnyx carrier api issues (#6615)
Fixes to the Telnyx carrier module

The best effort configuration setting did not work, My guess is that
this was written and then telnyx changed the api response and the
response to no longer included the flag “any_best_effort” so the old
code would never filter the best effort out. This improvement will now
pass the best effort flag to telnyx and remove the need to filter the
response.

Requests for TF numbers was failing as the request built and passed to
Telnyx was incorrect. It passed the TF NAP as `prefix` in the body and
it should be `nap` as there api doc states. I have also added the
option to filer TF numbers by NXX.

Also found that the crossbar quantity option was always overwritten to
500 so every time you requested 15 numbers through CB it requested 500
from telnyx resulting in longer query times. The fix was to only set
the default values if they were not set.