[4.3] HELP-15749: start kazoo_number_manager before kazoo_events (#6607)
When blackhole runs stand-alone, it relies on kazoo_events for the
kz_hooks when subscriptions for channel events are made. When the
event is a CHANNEL_CREATE from offnet, and no account ID is
associated, kz_hooks_util uses knm_number to lookup the account
associated with the destination.

kazoo_events.app did not list kazoo_number_manager as a dependency;
thus knm was not started and number lookups failed since the knm cache
wasn't started.

Updating kazoo_events.app to its proper dependencies resolves the issue.