read exisiting account document to get tree when importing enabled field
When trying to disable an account via the Account Manager app, the account is not disabled.

This bug was introduce when account creation/deletion was moved to core.
Before that move, we properly load and merge the request document with doc from database
so we had the `pvt_tree` field when we were checking ig the requestor is in
account's tree.
But this got moved into after validation and in `post/1,2` function and so we
don't set enabled anymore.