PISTON-1123: Fixes voicemail forward confirmation.
The `cf_voicemail:forward_message/6` function has been refactored to consistently execute the forwarding operation as an asynchronous task and to play the confirmation of the operation synchronously. The previous flow of control was inconsistent, depending on configuration. In some circumstances, the forwarding operation was performed asynchronously, but this resulted in no audible confirmation. In other circumstances, forwarding was performed synchronously.